# Improve performance of alignment