```
Benchmark cobot-bench: RUNNING...
benchmarking main/Local alignment
time                 2.816 s    (2.057 s .. 4.356 s)
                     0.963 R²   (0.945 R² .. 1.000 R²)
mean                 2.250 s    (2.086 s .. 2.541 s)
std dev              291.8 ms   (11.37 ms .. 350.1 ms)
variance introduced by outliers: 24% (moderately inflated)

benchmarking main/Global alignment
time                 1.703 s    (986.1 ms .. 2.311 s)
                     0.980 R²   (0.925 R² .. 1.000 R²)
mean                 1.602 s    (1.504 s .. 1.734 s)
std dev              140.0 ms   (39.78 ms .. 187.7 ms)
variance introduced by outliers: 22% (moderately inflated)

benchmarking main/Semiglobal alignment
time                 1.640 s    (412.3 ms .. 2.519 s)
                     0.941 R²   (0.795 R² .. 1.000 R²)
mean                 1.992 s    (1.703 s .. 2.113 s)
std dev              219.5 ms   (27.29 ms .. 280.0 ms)
variance introduced by outliers: 23% (moderately inflated)

benchmarking main/Local alignment with affine gap
time                 2.487 s    (1.101 s .. 3.042 s)
                     0.964 R²   (0.871 R² .. 1.000 R²)
mean                 2.370 s    (2.221 s .. 2.621 s)
std dev              239.1 ms   (2.451 ms .. 290.8 ms)
variance introduced by outliers: 23% (moderately inflated)

benchmarking main/Global alignment with affine gap
time                 1.728 s    (1.212 s .. 2.364 s)
                     0.978 R²   (0.973 R² .. 1.000 R²)
mean                 1.796 s    (1.643 s .. 2.012 s)
std dev              204.6 ms   (1.118 ms .. 250.2 ms)
variance introduced by outliers: 23% (moderately inflated)

benchmarking main/Semiglobal alignment with affine gap
time                 1.702 s    (1.540 s .. 2.006 s)
                     0.996 R²   (0.994 R² .. 1.000 R²)
mean                 1.700 s    (1.647 s .. 1.752 s)
std dev              63.78 ms   (40.49 ms .. 79.38 ms)
variance introduced by outliers: 19% (moderately inflated)

Benchmark cobot-bench: FINISH
```